refactor: Organizando as pastas componentes
