+++ components/camera/CameraPositionHelper.tsx
