+++ components/scene/PictureFrame.tsx
