+++ components/sequences/EndGameSequence.tsx
