+++ components/sequences/IntroNarrativa.tsx
